+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - root_pid.py accepts graceful termination timeout argument
+- added option --utc to log messages with UTC timestamp instead of local time
 
 ### Changed
 - refactored root_pid.py to depend only on python3-minimal